# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Glen ID Platform API
  description: |
    Glen ID Platformは、統合認証・認可基盤です。複数の認証方式（パスワード、WebAuthn、ソーシャルログイン）をサポートし、
    外部サービスに対してシングルサインオン（SSO）機能を提供します。
    
    ## 主要機能
    - ユーザー管理（登録、認証、プロフィール管理）
    - JWT ベースの認証・認可
    - WebAuthn（生体認証、セキュリティキー）
    - ソーシャルログイン（Google、GitHub、Discord）
    - API キー管理
    - 外部サービス統合
    
    ## 認証フロー
    1. **パスワード認証**: 従来のユーザー名/パスワード
    2. **WebAuthn認証**: 生体認証やセキュリティキーによるパスワードレス認証
    3. **ソーシャル認証**: OAuth2によるソーシャルプロバイダー連携
    
    ## セキュリティ
    - JWT トークン（アクセストークン + リフレッシュトークン）
    - PKCE（Proof Key for Code Exchange）対応
    - CORS設定による適切なアクセス制御
    - レート制限
  version: "1.0.0"
  contact:
    name: Glen ID Platform
    url: https://github.com/dqx0/glen
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.glen.dqx0.com/api/v1
    description: 本番環境
  - url: https://staging-api.glen.dqx0.com/api/v1
    description: ステージング環境
  - url: http://localhost:8080/api/v1
    description: 開発環境
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: ユーザーログイン
      description: ユーザー名/パスワードによる認証を行い、JWTトークンを発行します
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: "user@example.com"
              password: "securepassword123"
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /auth/logout:
    post:
      tags: [auth]
      summary: ログアウト
      description: 現在のセッションを終了し、トークンを無効化します
      responses:
        '200':
          description: ログアウト成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWTアクセストークンによる認証
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: APIキーによる認証
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "操作が正常に完了しました"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "リクエストに不正な値が含まれています"
        details:
          type: object
          additionalProperties: true

    # ===== 認証関連 =====
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "securepassword123"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWTアクセストークン
        refresh_token:
          type: string
          description: リフレッシュトークン
        token_type:
          type: string
          example: "Bearer"
        expires_in:
          type: integer
          description: アクセストークンの有効期限（秒）
          example: 3600
        user:
          $ref: '#/components/schemas/User'
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: 不正なリクエスト
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 認証が必要
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: レート制限に達しました
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'